requirements.txt
# duckdb
duckdb==1.0.0

# git
GitPython==3.1.43
